refactor(actor/system/vm): first pass at group support.
This commit implements early stage group support by allowing actors
to be assigned to a group when spawned.

For now, the only thing you can do with a group is kill all the actors
in it. Groups can be used in place of addresses by prefixing them with
the '$' symbol.

In the near future, groups will also allow messsages to be sent to all actors
in a group.